Fix for negative sizes interpreted as positive for endianness method.

+++ src/mutibs.rs
@@ -876,7 +876,7 @@ impl Mutibs {
         length: i64,
         endianness: Option<Endianness>,
     ) -> PyResult<Self> {
-        let is_little_endian = Endianness::is_little_endian(endianness, length as usize)?;
+        let is_little_endian = Endianness::is_little_endian(endianness, length)?;
         let bv = bv_from_u128(u, length, is_little_endian)?;
         Ok(Mutibs::from_bv(bv))
     }
@@ -952,7 +952,7 @@ impl Mutibs {
         length: i64,
         endianness: Option<Endianness>,
     ) -> PyResult<Self> {
-        let is_little_endian = Endianness::is_little_endian(endianness, length as usize)?;
+        let is_little_endian = Endianness::is_little_endian(endianness, length)?;
         let bv = bv_from_i128(i, length, is_little_endian)?;
         Ok(Mutibs::from_bv(bv))
     }
@@ -1026,7 +1026,7 @@ impl Mutibs {
         length: i64,
         endianness: Option<Endianness>,
     ) -> PyResult<Self> {
-        let is_little_endian = Endianness::is_little_endian(endianness, length as usize)?;
+        let is_little_endian = Endianness::is_little_endian(endianness, length)?;
         let bv = bv_from_f64(f, length, is_little_endian)?;
         Ok(Mutibs::from_bv(bv))
     }
